imageai/requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageai==2.1.5
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kafka-python==1.4.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mock==3.0.5
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
six==1.13.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.4.0
u-msgpack-python==2.5.2
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
